% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStepwise.R
\name{extractp}
\alias{extractp}
\title{Easily access p-values from linear models}
\usage{
extractp(pred, fit)
}
\arguments{
\item{pred}{the predictor from the linear model whose p-value is desired}

\item{fit}{a linear model of type "lm" containing the desired predictor}
}
\value{
the p-value for the desired predictor from the linear model
}
\description{
This function gives the user easy access to p-values for specific predictors from a linear model.  It is mainly used
to easily pass p-values to other functions.
}
\author{
Cory Langille <lang1729@gmail.com>
}
\seealso{
\code{lm}
}

