% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStepwise.R
\name{fMaker}
\alias{fMaker}
\title{Create formulas from strings to be used in linear models}
\usage{
fMaker(pred, fitCurrent, add = T)
}
\arguments{
\item{pred}{the predictor to be added or removed from the current model}

\item{fitCurrent}{the current model to be updated}

\item{add}{by default adds the predictor to to the model.  add=F removes the predictor from the model}
}
\value{
the updated model of type "lm"
}
\description{
This function takes a string and a linear model and either adds the predictor to the model, or removes it.
}

