% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafshape.R
\docType{data}
\name{leafshape}
\alias{leafshape}
\title{Leaf data from DAAG dataset leafshape}
\format{Dataset with 9 variables}
\source{
R dataset leafshape::DAAG
}
\usage{
data(leafshape)
}
\description{
Leaf length, width and petiole measurements taken at various sites in Australia.
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}

