% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStepwise.R
\name{pStepwise}
\alias{pStepwise}
\title{Selects the best predictors for a linear model based on p-values}
\usage{
pStepwise(response, fullmodel, aEnter = 0.1, aRemove = 0.1)
}
\arguments{
\item{response}{the response variable of interest in the model}

\item{fullmodel}{a linear model containing all possible predictors.  Typically of the form lm(y~., data=data)}

\item{aEnter}{the threshold for adding new predictors, set to 0.1 by default}

\item{aRemove}{the threshold for removing predictors from the current model, set to 0.1 by default}
}
\value{
a linear model of type lm containing the "best" predictors
}
\description{
This function will create a linear model based on the p-values for each predictor.
}
\author{
Cory Langille <lang1729@gmail.com>
}
\seealso{
extractp, stepfwd, stepbwd, fMaker
}

